back to original error

+++ domain/src/main/java/me/rooshi/domain/manager/KeyManager.kt
@@ -4,5 +4,5 @@ interface KeyManager {
 
     fun reset()
 
-    fun newId(): Int
+    fun newId(): Long
 }
